+++ SharePointFramework/shared-library/src/components/Toolbar/Toolbar.tsx
@@ -6,7 +6,7 @@ import {
 } from '@fluentui/react-components'
 import { FilterPanel } from '../FilterPanel'
 import React, { FC } from 'react'
-import { renderToolbarItem } from './renderToolbarItem'
+import { useToolbarItemRender } from './useToolbarItemRender'
 import styles from './Toolbar.module.scss'
 import { IToolbarProps } from './types'
 import strings from 'SharedLibraryStrings'
@@ -14,6 +14,7 @@ import { customLightTheme } from '../../util'
 
 export const Toolbar: FC<IToolbarProps> = (props) => {
   const fluentProviderId = useId('fp-toolbar')
+  const { renderToolbarItem } = useToolbarItemRender()
   return (
     <IdPrefixProvider value={fluentProviderId}>
       <FluentProvider theme={customLightTheme} className={styles.root}>
