--- SharePointFramework/shared-library/src/components/Toolbar/renderToolbarItem.ts
@@ -1,15 +1,0 @@
-import { ListMenuItem } from './types'
-import { renderMenu } from './renderMenu'
-import { renderToolbarButton } from './renderToolbarButton'
-
-/**
- * Renders a toolbar item.
- *
- * @param item - The item to render.
- *
- * @returns The rendered toolbar item.
- */
-export function renderToolbarItem(item: ListMenuItem) {
-  if (item.items) return renderMenu(item)
-  else return renderToolbarButton(item)
-}
